Move conversation and message modles to models/messaging/. Move conversations and messages controllers to controllers/messages.
